change project cmake structure

+++ src/GBufferPass.cpp
@@ -1,6 +1,5 @@
 #include "GBufferPass.h"
 #include "shader/HostDevice.h"
-#include "core/DebugUtils.h"
 
 GBufferPass::GBufferPass(
 	const zvk::Context* ctx, vk::Extent2D extent, const Resource& resource, vk::ImageLayout outLayout
@@ -47,7 +46,7 @@ void GBufferPass::render(vk::CommandBuffer cmd, vk::Extent2D extent, uint32_t in
 	cmd.bindVertexBuffers(0, param.vertexBuffer, vk::DeviceSize(0));
 	cmd.bindIndexBuffer(param.indexBuffer, 0, vk::IndexType::eUint32);
 
-	cmd.setViewport(0, vk::Viewport(0.0f, 0.0f, extent.width, extent.height, 0.0f, 1.0f));
+	cmd.setViewport(0, vk::Viewport(0.0f, 0.0f, static_cast<float>(extent.width), static_cast<float>(extent.height), 0.0f, 1.0f));
 	cmd.setScissor(0, vk::Rect2D({ 0, 0 }, extent));
 
 	for (uint32_t i = 0; i < param.count; i++) {
